django-webserver/pyproject.toml
[project]
name = "django-webserver"
version = "0.1.0"
description = "Django + Resonate example webserver"
authors = [
    { name = "ResonateHQ", email = "contact@resonatehq.io" }
]
dependencies = [
    "django>=4.2.16",
    "resonate-sdk>=0.4.12",
]
readme = "README.md"
requires-python = ">= 3.9"

[project.scripts]
"django-webserver" = "manage:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/django_webserver"]
